split loginFragment into multiple files
